PMGR: Finalize extraction of is_os_compatible into octoprint.util.platform
